Move categoryId Validator to common decorator
