+++ server/src/services/category/decorator/index.ts
@@ -1,5 +1,4 @@
 export * from './createCategory.decorator'
-export * from './validateIdParam.decorator'
 export * from './validator'
 export * from './deletedCheck.decorator'
 export * from './updateCategory.decorator'
