+++ server/src/services/category/decorator/validator/index.ts
@@ -1,4 +1,3 @@
 export * from './createCategory.validator'
-export * from './categoryIdParam.validator'
 export * from './deleteCheck.validator'
 export * from './updateCategory.validator'
