+++ server/src/services/common/decorator/index.ts
@@ -1,0 +1,1 @@
+export * from './validateIdParam.decorator'
