+++ server/src/services/common/types/params.type.ts
@@ -1,0 +1,5 @@
+import { tags } from 'typia'
+
+export interface CategoryIdParamsDto {
+  categoryId: string & tags.Format<'uuid'>
+}
