+++ server/src/services/common/validator/index.ts
@@ -1,0 +1,1 @@
+export * from './categoryIdParam.validator'
